tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.quadraticCurveTo.nonfinite.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.quadraticCurveTo.nonfinite</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.quadraticCurveTo.nonfinite</h1>
<p class="desc">quadraticCurveTo() with Infinity/NaN is ignored</p>


<script>
var t = async_test("quadraticCurveTo() with Infinity/NaN is ignored");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.moveTo(0, 0);
  ctx.lineTo(100, 0);
  ctx.quadraticCurveTo(Infinity, 50, 0, 50);
  ctx.quadraticCurveTo(-Infinity, 50, 0, 50);
  ctx.quadraticCurveTo(NaN, 50, 0, 50);
  ctx.quadraticCurveTo(0, Infinity, 0, 50);
  ctx.quadraticCurveTo(0, -Infinity, 0, 50);
  ctx.quadraticCurveTo(0, NaN, 0, 50);
  ctx.quadraticCurveTo(0, 50, Infinity, 50);
  ctx.quadraticCurveTo(0, 50, -Infinity, 50);
  ctx.quadraticCurveTo(0, 50, NaN, 50);
  ctx.quadraticCurveTo(0, 50, 0, Infinity);
  ctx.quadraticCurveTo(0, 50, 0, -Infinity);
  ctx.quadraticCurveTo(0, 50, 0, NaN);
  ctx.quadraticCurveTo(Infinity, Infinity, 0, 50);
  ctx.quadraticCurveTo(Infinity, Infinity, Infinity, 50);
  ctx.quadraticCurveTo(Infinity, Infinity, Infinity, Infinity);
  ctx.quadraticCurveTo(Infinity, Infinity, 0, Infinity);
  ctx.quadraticCurveTo(Infinity, 50, Infinity, 50);
  ctx.quadraticCurveTo(Infinity, 50, Infinity, Infinity);
  ctx.quadraticCurveTo(Infinity, 50, 0, Infinity);
  ctx.quadraticCurveTo(0, Infinity, Infinity, 50);
  ctx.quadraticCurveTo(0, Infinity, Infinity, Infinity);
  ctx.quadraticCurveTo(0, Infinity, 0, Infinity);
  ctx.quadraticCurveTo(0, 50, Infinity, Infinity);
  ctx.lineTo(100, 50);
  ctx.lineTo(0, 50);
  ctx.fillStyle = '#0f0';
  ctx.fill();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 90,45, 0,255,0,255);
  t.done();

});
</script>
